requirements.txt
webvtt-py
pyinstaller
TkinterDnD2
customtkinter